adds restart policy to docker container tasks
My droplet restarted for some reason, maybe the
result of auto updating packages. But once it
started up again, the docker containers for
pronouns and litestream didn't restart. Linkding
did restart because of the restart policy set
in the docker compose file. So I just added
the restart policy to the other tasks that create
docker containers.